requirements.txt
selenium
beautifulsoup4
pandas>=2.2.3
streamlit
python-dotenv
webdriver-manager
